@@ -32,6 +32,7 @@ import lombok.NonNull;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import net.devh.boot.grpc.client.security.CallCredentialsHelper;
+import org.factcast.core.DuplicateFactException;
 import org.factcast.core.Fact;
 import org.factcast.core.snap.Snapshot;
 import org.factcast.core.snap.SnapshotId;
@@ -159,13 +160,27 @@ public class GrpcFactStore implements FactStore {
   public void publish(@NonNull List<? extends Fact> factsToPublish) {
     runAndHandle(
         () -> {
-          log.trace("publishing {} facts to remote store", factsToPublish.size());
-          List<MSG_Fact> mf =
-              factsToPublish.stream().map(converter::toProto).collect(Collectors.toList());
-          MSG_Facts mfs = MSG_Facts.newBuilder().addAllFact(mf).build();
+          try {
+            log.trace("publishing {} facts to remote store", factsToPublish.size());
+            List<MSG_Fact> mf =
+                factsToPublish.stream().map(converter::toProto).collect(Collectors.toList());
+            MSG_Facts mfs = MSG_Facts.newBuilder().addAllFact(mf).build();
 
-          //noinspection ResultOfMethodCallIgnored
-          blockingStub.publish(mfs);
+            //noinspection ResultOfMethodCallIgnored
+            blockingStub.publish(mfs);
+          } catch (Exception e) {
+            RuntimeException decodedException = ClientExceptionHelper.from(e);
+            if (decodedException instanceof DuplicateFactException
+                && properties.isIgnoreDuplicateFacts()) {
+              // if it is only one, just skip
+              if (factsToPublish.size() > 1) {
+                //  fall back to publishing one by one
+                factsToPublish.forEach(f -> publish(Collections.singletonList(f)));
+              }
+            } else {
+              throw e;
+            }
+          }
         });
   }
 
@@ -179,13 +194,14 @@ public class GrpcFactStore implements FactStore {
         block.run();
         return;
       } catch (Exception e) {
-        if (resilience.shouldRetry(e)) {
-          log.warn("Temporary failure", e);
+        RuntimeException decodedException = ClientExceptionHelper.from(e);
+        if (resilience.shouldRetry(decodedException)) {
+          log.warn("Temporary failure", decodedException);
           log.info("Retry call to remote server");
           resilience.sleepForInterval();
           // continue and try next attempt
         } else {
-          throw ClientExceptionHelper.from(e);
+          throw decodedException;
         }
       }
     }
@@ -201,13 +217,14 @@ public class GrpcFactStore implements FactStore {
         T call = block.call();
         return call;
       } catch (Exception e) {
-        if (resilience.shouldRetry(e)) {
-          log.warn("Temporary failure", e);
+        RuntimeException decodedException = ClientExceptionHelper.from(e);
+        if (resilience.shouldRetry(decodedException)) {
+          log.warn("Temporary failure", decodedException);
           log.info("Retry call to remote server");
           resilience.sleepForInterval();
           // continue and try next attempt
         } else {
-          throw ClientExceptionHelper.from(e);
+          throw decodedException;
         }
       }
     }
