@@ -15,20 +15,21 @@
  */
 package org.factcast.client.grpc;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTimeout;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.*;
 
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
 import java.io.IOException;
 import java.time.Duration;
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.UUID;
-import java.util.concurrent.TimeoutException;
-import java.util.function.Consumer;
+import java.util.*;
+import java.util.concurrent.*;
+import java.util.function.*;
 import lombok.NonNull;
 import lombok.SneakyThrows;
 import org.factcast.client.grpc.FactCastGrpcClientProperties.ResilienceConfiguration;
@@ -42,8 +43,10 @@ import org.factcast.core.subscription.Subscription;
 import org.factcast.core.subscription.SubscriptionClosedException;
 import org.factcast.core.subscription.SubscriptionRequestTO;
 import org.factcast.core.subscription.observer.FactObserver;
-import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.extension.*;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -93,15 +96,15 @@ class ResilientGrpcSubscriptionTest {
         .thenThrow(TimeoutException.class)
         .then(x -> subscription);
 
-    assertThrows(TimeoutException.class, () -> uut.awaitComplete(51));
+    assertThrows(TimeoutException.class, () -> uut.awaitComplete(551));
 
     assertTimeout(
         Duration.ofMillis(1000),
         () -> {
-          assertThat(uut.awaitComplete(52)).isSameAs(uut);
+          assertThat(uut.awaitComplete(552)).isSameAs(uut);
         });
     // await call was passed
-    verify(subscription).awaitComplete(52);
+    verify(subscription).awaitComplete(552);
   }
 
   @Test
@@ -110,15 +113,15 @@ class ResilientGrpcSubscriptionTest {
         .thenThrow(TimeoutException.class)
         .then(x -> subscription);
 
-    assertThrows(TimeoutException.class, () -> uut.awaitCatchup(51));
+    assertThrows(TimeoutException.class, () -> uut.awaitCatchup(551));
 
     assertTimeout(
         Duration.ofMillis(1000),
         () -> {
-          assertThat(uut.awaitCatchup(52)).isSameAs(uut);
+          assertThat(uut.awaitCatchup(552)).isSameAs(uut);
         });
     // await call was passed
-    verify(subscription).awaitCatchup(52);
+    verify(subscription).awaitCatchup(552);
   }
 
   @Test
