@@ -15,15 +15,16 @@
  */
 package org.factcast.core.event;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.*;
 import lombok.NonNull;
 import org.factcast.factus.event.EventObject;
 import org.factcast.factus.event.EventSerializer;
 import org.factcast.factus.event.Specification;
-import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.extension.*;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -46,11 +47,31 @@ class EventConverterTest {
         .isInstanceOf(IllegalArgumentException.class);
   }
 
+  @Test
+  void acceptsNullValueInMeta() {
+    underTest.toFact(new E3());
+  }
+
+  @Test
+  void discoversType() {
+    assertThat(underTest.toFact(new E3()).type()).isEqualTo(E3.class.getSimpleName());
+  }
+
   @Specification(ns = "test", version = -99)
   static class E implements EventObject {
     @Override
     public Set<UUID> aggregateIds() {
       return new HashSet<>();
+    }
+  }
+
+  @Specification(ns = "test", version = 10)
+  static class E3 extends E {
+    @Override
+    public Map<String, String> additionalMetaMap() {
+      HashMap<String, String> m = new HashMap<String, String>();
+      m.put("foo", null);
+      return m;
     }
   }
 
