@@ -1,5 +1,5 @@
 /*
- * Copyright © 2017-2020 factcast.org
+ * Copyright © 2017-2022 factcast.org
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.factcast.core;
+package org.factcast.itests.factus.v3compat;
 
-import lombok.NonNull;
+import org.factcast.itests.factus.FactusVariousHandlerParametersDemoITest;
+import org.factcast.test.FactcastTestConfig;
 
-public class MaxRetryAttemptsExceededException extends RuntimeException {
-
-  private static final long serialVersionUID = 1L;
-
-  public MaxRetryAttemptsExceededException(@NonNull String msg) {
-    super(msg);
-  }
-}
+@FactcastTestConfig(factcastVersion = "0.3.9")
+public class FactusVariousHandlerParametersDemoITestCompatV3
+    extends FactusVariousHandlerParametersDemoITest {}
