@@ -15,7 +15,7 @@
  */
 package org.factcast.store.internal;
 
-import java.util.Random;
+import java.util.*;
 import lombok.AccessLevel;
 import lombok.Generated;
 import lombok.NonNull;
@@ -70,6 +70,7 @@ public class PgConstants {
 
   public static final String CHANNEL_FACT_INSERT = "fact_insert";
   public static final String CHANNEL_SCHEDULED_POLL = "scheduled-poll";
+  public static final String CHANNEL_BLACKLIST_CHANGE = "blacklist_change";
   public static final String CHANNEL_ROUNDTRIP =
       "roundtrip_channel_"
           + Math.abs(new Random().nextLong()); // using the pid lead to a sql exception
