@@ -17,7 +17,7 @@ package org.factcast.store.internal;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.eventbus.EventBus;
-import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.*;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,6 +25,7 @@ import org.factcast.core.subscription.*;
 import org.factcast.core.subscription.observer.FactObserver;
 import org.factcast.core.subscription.observer.FastForwardTarget;
 import org.factcast.store.internal.catchup.PgCatchupFactory;
+import org.factcast.store.internal.filter.PgBlacklist;
 import org.factcast.store.internal.query.PgFactIdToSerialMapper;
 import org.factcast.store.internal.query.PgLatestSerialFetcher;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -48,6 +49,7 @@ class PgSubscriptionFactory {
   final FactTransformersFactory transformersFactory;
   final FastForwardTarget target;
   final PgMetrics metrics;
+  final PgBlacklist blacklist;
 
   public Subscription subscribe(SubscriptionRequestTO req, FactObserver observer) {
     SubscriptionImpl subscription =
@@ -61,7 +63,8 @@ class PgSubscriptionFactory {
             fetcher,
             catchupFactory,
             target,
-            metrics);
+            metrics,
+            blacklist);
 
     // when closing the subscription, also close the PgFactStream
     subscription.onClose(pgsub::close);
