@@ -15,19 +15,21 @@
  */
 package org.factcast.store.internal.catchup;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.*;
 import lombok.NonNull;
 import org.factcast.core.subscription.SubscriptionImpl;
 import org.factcast.core.subscription.SubscriptionRequestTO;
 import org.factcast.store.internal.PgMetrics;
-import org.factcast.store.internal.PgPostQueryMatcher;
+import org.factcast.store.internal.filter.PgFactFilter;
+import org.factcast.store.internal.query.CurrentStatementHolder;
 
 public interface PgCatchupFactory {
 
   PgCatchup create(
       @NonNull SubscriptionRequestTO request,
-      @NonNull PgPostQueryMatcher postQueryMatcher,
+      @NonNull PgFactFilter postQueryMatcher,
       @NonNull SubscriptionImpl subscription,
       @NonNull AtomicLong serial,
-      @NonNull PgMetrics metrics);
+      @NonNull PgMetrics metrics,
+      @NonNull CurrentStatementHolder statementHolder);
 }
