@@ -17,7 +17,6 @@ package org.factcast.store.internal.catchup.tmppaged;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.util.List;
-import java.util.UUID;
 import java.util.concurrent.atomic.AtomicLong;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
@@ -28,10 +27,11 @@ import org.factcast.core.subscription.SubscriptionImpl;
 import org.factcast.core.subscription.SubscriptionRequestTO;
 import org.factcast.store.StoreConfigurationProperties;
 import org.factcast.store.internal.PgMetrics;
-import org.factcast.store.internal.PgPostQueryMatcher;
 import org.factcast.store.internal.StoreMetrics;
 import org.factcast.store.internal.catchup.PgCatchup;
+import org.factcast.store.internal.filter.PgFactFilter;
 import org.factcast.store.internal.listen.PgConnectionSupplier;
+import org.factcast.store.internal.query.CurrentStatementHolder;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.datasource.SingleConnectionDataSource;
 
@@ -42,10 +42,11 @@ public class PgTmpPagedCatchup implements PgCatchup {
   @NonNull final PgConnectionSupplier connectionSupplier;
   @NonNull final StoreConfigurationProperties props;
   @NonNull final SubscriptionRequestTO request;
-  @NonNull final PgPostQueryMatcher postQueryMatcher;
+  @NonNull final PgFactFilter filter;
   @NonNull final SubscriptionImpl subscription;
   @NonNull final AtomicLong serial;
   @NonNull final PgMetrics metrics;
+  @NonNull final CurrentStatementHolder statementHolder;
 
   @SneakyThrows
   @Override
@@ -57,6 +58,7 @@ public class PgTmpPagedCatchup implements PgCatchup {
       fetch(jdbc);
     } finally {
       ds.destroy();
+      statementHolder.statement(null);
     }
   }
 
@@ -65,27 +67,22 @@ public class PgTmpPagedCatchup implements PgCatchup {
     long factCounter = 0L;
     jdbc.execute("CREATE TEMPORARY TABLE catchup(ser bigint)");
 
-    PgCatchUpPrepare prep = new PgCatchUpPrepare(jdbc, request);
+    PgCatchUpPrepare prep = new PgCatchUpPrepare(jdbc, request, statementHolder);
     // first collect all the sers
     var numberOfFactsToCatchUp = prep.prepareCatchup(serial);
     // and AFTERWARDs create the inmem index
     jdbc.execute("CREATE INDEX catchup_tmp_idx1 ON catchup(ser ASC)"); // improves perf on sorting
 
-    var skipTesting = postQueryMatcher.canBeSkipped();
-
     if (numberOfFactsToCatchUp > 0) {
-      PgCatchUpFetchTmpPage fetch = new PgCatchUpFetchTmpPage(jdbc, props.getPageSize(), request);
+      PgCatchUpFetchTmpPage fetch =
+          new PgCatchUpFetchTmpPage(jdbc, props.getPageSize(), request, statementHolder);
       List<Fact> facts;
       do {
         facts = fetch.fetchFacts(serial);
-
         for (Fact f : facts) {
-          UUID factId = f.id();
-          if (skipTesting || postQueryMatcher.test(f)) {
+          if (filter.test(f)) {
             subscription.notifyElement(f);
             factCounter++;
-          } else {
-            log.trace("{} filtered id={}", request, factId);
           }
         }
       } while (!facts.isEmpty());
