@@ -17,10 +17,9 @@ package org.factcast.store.internal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
-import java.util.UUID;
+import java.util.*;
 import org.factcast.store.internal.query.PgLatestSerialFetcher;
 import org.factcast.store.test.IntegrationTest;
 import org.junit.jupiter.api.Test;
@@ -34,7 +33,7 @@ import org.springframework.test.context.jdbc.SqlConfig;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ContextConfiguration(classes = {PgTestConfiguration.class})
-@Sql(scripts = "/test_schema.sql", config = @SqlConfig(separator = "#"))
+@Sql(scripts = "/wipe.sql", config = @SqlConfig(separator = "#"))
 @ExtendWith(SpringExtension.class)
 @IntegrationTest
 public class PgLatestSerialFetcherTest {
@@ -57,7 +56,7 @@ public class PgLatestSerialFetcherTest {
             + PgConstants.COLUMN_PAYLOAD
             + ") VALUES('{\"id\":\""
             + UUID.randomUUID()
-            + "\"}','{}') ");
+            + "\", \"ns\":\"hups\"}','{}') ");
     assertEquals(1, uut.retrieveLatestSer());
     jdbcTemplate.execute(
         "INSERT INTO "
@@ -68,7 +67,7 @@ public class PgLatestSerialFetcherTest {
             + PgConstants.COLUMN_PAYLOAD
             + ") VALUES('{\"id\":\""
             + UUID.randomUUID()
-            + "\"}','{}') ");
+            + "\", \"ns\":\"hups\"}','{}') ");
     jdbcTemplate.execute(
         "INSERT INTO "
             + PgConstants.TABLE_FACT
@@ -78,7 +77,7 @@ public class PgLatestSerialFetcherTest {
             + PgConstants.COLUMN_PAYLOAD
             + ") VALUES('{\"id\":\""
             + UUID.randomUUID()
-            + "\"}','{}') ");
+            + "\", \"ns\":\"hups\"}','{}') ");
     assertEquals(3, uut.retrieveLatestSer());
   }
 
