@@ -23,20 +23,19 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.jdbc.SqlConfig;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.transaction.support.TransactionTemplate;
 
 @ContextConfiguration(classes = {PgTestConfiguration.class})
-@Sql(scripts = "/test_schema.sql", config = @SqlConfig(separator = "#"))
 @ExtendWith(SpringExtension.class)
 @ExtendWith(MockitoExtension.class)
 @IntegrationTest
 class PgTransformationStoreImplTest extends AbstractTransformationStoreTest {
-  @Autowired private JdbcTemplate tpl;
+  @Autowired private JdbcTemplate jdbcTpl;
+  @Autowired private TransactionTemplate txTpl;
 
   @Override
   protected TransformationStore createUUT() {
-    return new PgTransformationStoreImpl(tpl, registryMetrics);
+    return new PgTransformationStoreImpl(jdbcTpl, txTpl, registryMetrics);
   }
 }
